+++ lib/src/ng/token.dart
@@ -1,0 +1,4 @@
+import 'package:angular/angular.dart';
+import 'package:lawndart/lawndart.dart';
+
+const OpaqueToken<Store> settingsStore = OpaqueToken('settingsStore');