+++ web/CHANGELOG.md
@@ -9,6 +9,7 @@
 * added changelog
 * remember nickname when losing
 * start game on enter
+* store settings in browser
 
 ## Bugfixes
 * background was not rendered on mobile phones since 0.4.0, switched back to highp precision and disabled background rendering on devices not supporting highp precision in fragment shaders